Debug XDMF-writing functions.
Runtime errors seem to be gone.  Valgrind errors are all within the Xdmf libraries and thus not under my easy control.

Generated file can be read-in by ParaView.  4D dataset causes errors since ParaView expects 3D or less.  Spacing and origin values within the XDMF XML file are for 3D, regardless of the shape of the actual data array.  Not sure that the values are correct for lower-dimensional data, though.