Add USE_VTK flags & improve CMakeLists configuration.
Attempting to make compilation possible without VTK support (will sacrifice some functionality of UniformVolume, but that's a decision left to the user).  Compilation without VTK not tested yet.

Made general improvements to CMakeLists.txt in order to improve build process of other projects.  In particular, this project is no longer compiled as a library.  Instead, it is intended to be included in the source of other projects which use it, either by an actual copy or a symbolic link from the project source to this directory (make this directory appear as a subdirectory of the project).